Move Amount impl into a submodule
